revert most cast issues reported by buggy gcc 3.4 snapshot of this week. Checked with gcc 3.4 from cvs of today which does not report any problems anymore.
git-svn-id: http://root.cern.ch/svn/root/trunk@7147 27541ba8-7e3a-0410-8455-c3a389f83636

+++ zip/src/Inflate.c
@@ -1176,7 +1176,7 @@ void R__unzip(int *srcsize, uch *src, int *tgtsize, uch *tgt, int *irep)
     There are some rare cases when a few more bytes are required */
   if (obufptr - tgt > *tgtsize) {
     fprintf(stderr,"R__unzip: discrepancy (%d) with initial size: %ld, tgtsize=%d\n",
-            obufptr - tgt,isize,(int)(*tgtsize));
+            obufptr - tgt,isize,*tgtsize);
     *irep = obufptr - tgt;
     return;
   }
